% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_find.R
\name{actively_find_funcs}
\alias{actively_find_funcs}
\title{Try harder to find functions that are called}
\usage{
actively_find_funcs(map, funcs)
}
\arguments{
\item{map}{Function map.}

\item{funcs}{Character vector, functions to find.}
}
\description{
We try to load imported and depended packages
and search their namespaces as well.
}
\details{
For this we need to install the package to a temporary
directory, and load it. If this is not possible, a warning
is given.

We asume that the trivial search in imported packages
and base packages did not work, so we don't try it again.
}
\keyword{internal}

