% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces.R
\name{add_namespaces}
\alias{add_namespaces}
\title{Add namespace information to a map of an R package}
\usage{
add_namespaces(map)
}
\arguments{
\item{map}{Function map.}
}
\value{
The annotated map.
}
\description{
Try to find out where the called functions live,
and add this information to the map.
}
\keyword{internal}

