% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRscript.R
\name{analyse.do.call.pattern}
\alias{analyse.do.call.pattern}
\title{analyse.do.call.pattern}
\usage{
analyse.do.call.pattern(e)
}
\arguments{
\item{e}{an expression or list of expressions}
}
\value{
list of expressions
}
\description{
This function helps user to add additional edges to the final functional Map
}
\details{
For \code{do.call} pattern is like

\code{do.call(sin, argList)}

it can be analyzed staticly, without knowning run-time information, hence it can be converted to a normal function call
}
\examples{
\dontrun{
      analyse.do.call.pattern( { 
          quote(do.call('myfun', list('arg1','arg2','arg3')) )
      })
      re <- analyse.do.call.pattern( { 
          quote(do.call(myfun, list(arg1,arg2,arg3)) )
      })
      convertToCharacter(re) }
}

