% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRscript.R
\name{analyse.external.call.pattern}
\alias{analyse.external.call.pattern}
\title{analyse.external.call.pattern}
\usage{
analyse.external.call.pattern(e)
}
\arguments{
\item{e}{expression}
}
\value{
list of experssions
}
\description{
match \code{.C}, \code{.Fortran} and \code{.Call}
}
\details{
A global option \code{add.prefix.for.external.call} will add \code{C_}, \code{FORTRAN_} and \code{External_} prefix or not
}
\examples{
\dontrun{
    analyse.external.call.pattern( quote( .C('classRF') ) )
    analyse.external.call.pattern( quote( .C(classRF) ) )
}
}

