% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticTestPattern.R
\name{annotate}
\alias{annotate}
\title{Annotate code with keywords}
\usage{
annotate(funcs = NULL, paths = NULL)
}
\arguments{
\item{funcs}{A character vector of function names to annotate.}

\item{paths}{A character vector of R files or folders with R files
to annotate.}
}
\value{
A table object (named integer vector) of all functions
}
\description{
Assess the code by finding out what kind of statistical methods
it uses.
}
\examples{
annotate("lm")

f <- function() {
  lm(Sepal.Width ~ Sepal.Length, data = iris)
}
annotate("f")

## Multiplicity of terms, 'nonlinear' appears twice
annotate(c("nlm", "nls"))
}

