% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticTestPattern.R
\name{annotate_functions}
\alias{annotate_functions}
\title{Annotate functions}
\usage{
annotate_functions(funcs)
}
\arguments{
\item{funcs}{A character vector of function names to annotate.}
}
\value{
A character vector of terms.
}
\description{
Annotate functions
}

