% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticTestPattern.R
\name{annotate_paths}
\alias{annotate_paths}
\title{Annotate all functions in R scripts or folders of scripts}
\usage{
annotate_paths(paths)
}
\arguments{
\item{paths}{A character vector of R files or folders with R files
to annotate.}
}
\value{
A character vector of terms.
}
\description{
Annotate all functions in R scripts or folders of scripts
}
\keyword{internal}

