% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverplot.R
\name{as_riverplot.function_map_rpackage}
\alias{as_riverplot.function_map_rpackage}
\title{Sankey diagram of an R package call graph}
\usage{
\method{as_riverplot}{function_map_rpackage}(x, ...)
}
\arguments{
\item{x}{The function map.}

\item{...}{Additional arguments to pass to the \code{makeRiver}
function in the \code{riverplot} package.}
}
\description{
This is currently the same as the basic method.
}

