% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverplot.R
\name{as_riverplot}
\alias{as_riverplot}
\title{Convert a function map to a Sankey plot (a \code{riverplot} object)}
\usage{
as_riverplot(x, ...)
}
\arguments{
\item{x}{The function map.}

\item{...}{Additional arguments to pass to the \code{makeRiver}
function in the \code{riverplot} package.}
}
\value{
A \code{riverplot} object.
}
\description{
A \code{riverplot} object describes a Sankey diagram, and
can be plotted via the \code{riverplot} package.
}

