% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticTestPattern.R
\name{assess.code}
\alias{assess.code}
\title{assess.code}
\usage{
assess.code(v, db = .db.stats)
}
\arguments{
\item{v}{vector of functions used}

\item{db}{the database of knowledge we used}
}
\value{
the external score of the function based directly on the statistical functions in db
}
\description{
assess the code by look for what kind of statistical routine used
}
\examples{
\dontrun{
   code0 <- 'ff <- function() {mod1 <- lm(Species ~ . , data=iris); summary(mod1)}' 
   cat(code0, file=(f0<-tempfile()))
   ret <- parseRscript(f0)
     assess.code( ret[['ff']] )
}
}

