% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticTestPattern.R
\name{assessRfolder}
\alias{assessRfolder}
\title{assessRfolder}
\usage{
assessRfolder(filepath, pattern = "\\\\.[R|r|S|s|Q|q]", db = .db.stats)
}
\arguments{
\item{filepath}{target path to be analysed}

\item{pattern}{file pattern to be analysed}

\item{db}{reference database , see \code{\link{assess.code}}}
}
\value{
list of functions with non-zero scores on statistical functions in \code{db}
}
\description{
analysis an folder of r source files using the \code{parseRscript} and \code{assess.code}
}

