% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_find.R
\name{attach_q}
\alias{attach_q}
\title{Attach a namespace quietly}
\usage{
attach_q(x)
}
\arguments{
\item{x}{Package name.}
}
\value{
The attached package env (from `attachNamespace`), invisibly.
}
\description{
Attach a namespace quietly
}
\keyword{internal}

