% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{bfs}
\alias{bfs}
\title{Perform a breadth first search (BFS) on a graph}
\usage{
bfs(graph, seeds)
}
\arguments{
\item{graph}{The input graph, an adjacency list.}

\item{seeds}{The seed vertices, a character vector.}
}
\value{
A character vector of vertex names, the visited vertices
  in the order of their visit.
}
\description{
Perform a BFS on a graph, from a set of seed vertices.
Once all vertices reachable from the seeds are visited,
the search terminates.
}
\keyword{internal}

