% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverplot.R
\name{calculate_x_pos}
\alias{calculate_x_pos}
\title{Calculate the X position of the functions on a Sankey plot
The network must be a DAG (directed, acyclic graph) to be able to
perform this. (Loop edges are OK, these are removed.)}
\usage{
calculate_x_pos(map)
}
\arguments{
\item{map}{The function map.}
}
\value{
A character vector, all functions in the function map,
in the order of topological sorting.
}
\description{
If the function map is not a DAG, then an error is reported.
}
\details{
We create the topological sort of the network first, and then
simply take all vertices in that order. Each vertex is assigned
a level one higher than the highest level of it's predecessors.
When no predecessors, level zero is assigned.
}

