% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pkg_dir}
\alias{check_pkg_dir}
\title{Check if a path is an R package root directory}
\usage{
check_pkg_dir(path = ".")
}
\arguments{
\item{path}{Path to the alleged package root.}
}
\description{
Check if a path is an R package root directory
}
\details{
It must have: \itemize{
  \item `DESCRIPTION`
  \item `NAMESPACE`
  \item A non-empty `R` folder.
}
}
\keyword{internal}

