% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_prefix}
\alias{common_prefix}
\title{The longest prefix of both lists}
\usage{
common_prefix(x, y)
}
\arguments{
\item{x}{First list.}

\item{y}{Second list.}
}
\value{
List, longest common prefix of both.
}
\description{
The longest prefix of both lists
}
\keyword{internal}

