% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRscript.R
\name{convertToCharacter}
\alias{convertToCharacter}
\title{convertToCharacter}
\usage{
convertToCharacter(L)
}
\arguments{
\item{L}{input list.}
}
\value{
character list representing input \code{L}
}
\description{
Convert a list names and characters to a character vector.
}

