% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{create_function_map}
\alias{create_function_map}
\title{Create a function_map object from a parsed result}
\usage{
create_function_map(data, rfile = NULL, rpath = NULL, rfilepattern = NULL,
  include_base = NULL, class = NULL)
}
\arguments{
\item{data}{A named list. Names are function names,
list entries are the functions they call. Make sure you include
multiplicity.}
}
\value{
A function_map object.
}
\description{
Create a function_map object from a parsed result
}

