% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetwork.R
\name{createNetwork}
\alias{createNetwork}
\title{Create the network object for the map of functions}
\usage{
createNetwork(funlist, omitpattern = "^\\\\.|\%", rootfunc = NULL,
  transpose = FALSE, returnmatrix = FALSE)
}
\arguments{
\item{funlist}{A list of functiions structure. The result of the function \code{\link{parseRfolder}}.}

\item{omitpattern}{An regular expression. The function names which match the regular expression will be omitted.}

\item{rootfunc}{A function name which will be used as root node. Only it's children will be returned. NULL for the whole (might be disconnected) network}

\item{transpose}{Whether to transpose the matrix of network structure.}

\item{returnmatrix}{Whether to return a matrix rather than the default network object.}
}
\value{
A network object or matrix.
}
\description{
Create the network object for the map of functions
}
\details{
If return a matrix, the row means the functions be invoked by the function of rowname, 1 means 'invoked'.
}
\examples{
\dontrun{
lfun <- parseRfolder(system.file("examples", "R", package = "functionMap"))
createNetwork(lfun)
}

}
\author{
Mango Solutions
}

