% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{deps}
\alias{deps}
\title{Dependencies of all functions}
\usage{
deps(map, multiples = FALSE)
}
\arguments{
\item{map}{Function map.}

\item{multiples}{Whether to keep multiple function calls to
the same function, i.e. duplicates in the list.}
}
\value{
Named list of character vectors.
}
\description{
Dependencies of all functions
}

