% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{dfs.matrix.travel}
\alias{dfs.matrix.travel}
\title{dfs.matrix.travel}
\usage{
dfs.matrix.travel(n, v, direction = "forward")
}
\arguments{
\item{n}{adj matrix}

\item{v}{index from which we start, missing then searhing all connected component}

\item{direction}{forward, backward or bidirection}
}
\value{
traveled vertexes indices from \code{v} or a list of connected components (missing \code{v})
}
\description{
Depth-First travel a graph in a matrix representation
}

