% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-call.R
\name{do_call_globals}
\alias{do_call_globals}
\title{Find the global functions in \code{do.call} patterns}
\usage{
do_call_globals(fun)
}
\arguments{
\item{fun}{A function object.}
}
\value{
Character vector of globals from \code{do.call} calls.
Note that this only includes globals that are speficied
in the \code{do.call} call as a character. Globals that are passed
to \code{do.call} as function objects, are found by the standard
\code{findGlobals} already.
}
\description{
Find the global functions in \code{do.call} patterns
}

