% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-frame.R
\name{edge_data_frame}
\alias{edge_data_frame}
\title{Extract edge (= function call) metadata from the call graph}
\usage{
edge_data_frame(map)
}
\arguments{
\item{map}{Function map.}
}
\value{
Data frame, one row for each edge.
}
\description{
Extract edge (= function call) metadata from the call graph
}
\keyword{internal}

