% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{edge_df}
\alias{edge_df}
\title{Extract data about the edges (function calls) in a function map}
\usage{
edge_df(map)
}
\arguments{
\item{map}{The function map.}
}
\value{
A data frame with columns:
  \item{from}{Calling function.}
  \item{to}{Called function.}
  \item{type}{Call type.}
  \item{line}{Line number.}
  \item{col1}{Starting column number.}
  \item{col2}{Ending column number.}
}
\description{
Extract data about the edges (function calls) in a function map
}

