% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{edgelist.from.rscript}
\alias{edgelist.from.rscript}
\title{edgelist.from.rscript}
\usage{
edgelist.from.rscript(rfile)
}
\arguments{
\item{rfile}{R script name}
}
\value{
The return value is a data.frame containing 5 columns.
\describe{
 \item{tails}{the caller}
 \item{heads}{the callee, and in the graph we will draw an arrow from tail pointing to head}
 \item{category}{currently can be\describe{ 
     \item{normal}{an ordinary function call}
     \item{do.call}{invoked using do.call mechanism}
     \item{native.call}{a native C or FORTRAN function, invoked by .C, .Fortran, .Call or .External}
     \item{eval.call}{invoked by \code{eval}, typically might be a variable or expression, rather than a existing function}
 }}
 \item{weights}{how many times the callee appearing in the caller}
 \item{rfile}{where this relation is parsed from}
}
}
\description{
Difference between this function and \code{\link{parseRscript}} is this one will give more information
of the edge and the node.
}
\examples{
\dontrun{
 rfile <- system.file("examples", "R", "func.R", package = "functionMap")
 edgelist.from.rscript(rfile)
}
}

