% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{edgelist.from.SASfolder}
\alias{edgelist.from.SASfolder}
\title{edgelist.from.SASfolder}
\usage{
edgelist.from.SASfolder(base.path, pattern = "\\\\.[Ss][Aa][Ss]$")
}
\arguments{
\item{base.path}{root path}

\item{pattern}{extension name pattern for the file to be parsed}
}
\value{
edgelist object(a data.fram)
}
\description{
Similar to \code{\link{edgelist.from.rpackage}}
}
\examples{
\dontrun{
     plot(eForce(network.from.edgelist(edgelist.from.SASfolder('.'))))
}
}

