% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{edgelist.from.SASscript}
\alias{edgelist.from.SASscript}
\title{edgelist.from.SASscript}
\usage{
edgelist.from.SASscript(sasfile)
}
\arguments{
\item{sasfile}{SAS filename}
}
\value{
data.frame with same structure as the output of \code{edgelist.from.rscript}
}
\description{
Similar to \code{\link{edgelist.from.rscript}}, create edgelist data.frame from a single SAS script
}
\examples{
\dontrun{
     plot(eForce(network.from.edgelist(edgelist.from.SASscript('Acorda.sas'))))
}
}

