% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{export_map}
\alias{export_map}
\title{Export a map to a JSON file}
\usage{
export_map(map, file, pretty = TRUE)
}
\arguments{
\item{map}{The function map.}

\item{file}{The file or connection to write to.}

\item{pretty}{Whether to pretty print the JSON file.}
}
\description{
Export a map to a JSON file
}

