% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{external_calls}
\alias{external_calls}
\title{Find calls to external functions}
\usage{
external_calls(fun, multiples = FALSE)
}
\arguments{
\item{fun}{Function object.}

\item{multiples}{Whether to include functions as many
times as they are called.}
}
\value{
Character vector of external function names,
  prefixed according to the call type.
}
\description{
It supports `.C`, `.Call`, `.Fortran` and `.External` calls.
}

