% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpS4Methods.R
\name{extract.S4.defn}
\alias{extract.S4.defn}
\title{extract.S4.defn}
\usage{
extract.S4.defn(srclist, path, single = TRUE, export.other.defn = TRUE)
}
\arguments{
\item{srclist}{source list}

\item{path}{alternatively, we can specify a path and use all R sources in it}

\item{single}{if we should paste the multiple characters into a single one}

\item{export.other.defn}{export other function definitions}
}
\value{
text include only S4 related definitions 
}
\description{
Extract S4 class and methods definition
}
\examples{
\dontrun{

     cat(extract.S4.defn(path=system.file("examples", "R", package = "functionMap")))
}
}

