% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_if_needed}
\alias{extract_if_needed}
\title{Extract a package tarball}
\usage{
extract_if_needed(path)
}
\arguments{
\item{path}{Path to a package tarball, or a package directory.}
}
\value{
Path to the temporary directory (or the supplied `path`
  if it was already a directory.
}
\description{
The package is extracted in a temporary directory.
}
\details{
If `path` is a directory, then we do nothing.
}
\keyword{internal}

