% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_find.R
\name{find_function}
\alias{find_function}
\title{Find a function in an environment or in its parent chain}
\usage{
find_function(func, env = parent.env())
}
\arguments{
\item{func}{Character scalar, function name.}

\item{env}{Environment to start the search from.}
}
\value{
Name of the environment of \code{func}, or
  \code{NA_character_} if it cannot be found.
}
\description{
Find a function in an environment or in its parent chain
}

