% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{find_globals}
\alias{find_globals}
\title{Find all (non-local) function calls in a function}
\usage{
find_globals(func, row, multiples = FALSE)
}
\arguments{
\item{func}{Function object.}

\item{row}{The row of the function in the parse data.}

\item{multiples}{Whether to keep multiplicity.}
}
\value{
Character vector of function calls.
}
\description{
Find all (non-local) function calls in a function
}
\keyword{internal}

