% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{find_globals_multiple}
\alias{find_globals_multiple}
\title{Find all (non-local) function calls in a function, keep multiple calls}
\usage{
find_globals_multiple(func)
}
\arguments{
\item{func}{Function object.}
}
\value{
Character vector of function calls.
}
\description{
Find all (non-local) function calls in a function, keep multiple calls
}
\keyword{internal}

