% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_in_named_list}
\alias{find_in_named_list}
\title{Find an element within entries of a named list}
\usage{
find_in_named_list(list, elem)
}
\arguments{
\item{list}{List.}

\item{elem}{Element.}
}
\value{
Name of the list entry in which `elem` was
  found, or `NA_character_` if it was not found.
}
\description{
Find an element within entries of a named list
}
\keyword{internal}

