% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-arguments.R
\name{func_arg_globals}
\alias{func_arg_globals}
\title{Find the global functions that are used as arguments}
\usage{
func_arg_globals(fun, multiples = FALSE)
}
\arguments{
\item{fun}{A function object.}

\item{multiples}{Whether to include functions as many
times as they are called.}
}
\value{
Character vector of globals from these calls.
}
\description{
In \code{do.call}, \code{apply}, \code{lapply}, \code{sapply},
\code{vapply}, \code{tapply}, \code{mapply}, \code{eapply},
\code{rapply} functions.
}
\keyword{internal}

