% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionMapTests.R
\name{functionMapTests}
\alias{functionMapTests}
\title{Run package regression tests}
\usage{
functionMapTests(unitTestPath = system.file(package = "functionMap",
  "unittests"), printTestProtocol = FALSE,
  printTestFile = "reportingunit.html")
}
\arguments{
\item{unitTestPath}{Path where the scripts are located.  By default, will use the package installation directory.}

\item{printTestProtocol}{Logical flag.  Should HTML reports be produced?}

\item{printTestFile}{The name of the HTML reports.}
}
\value{
The results of executing the test suites via the function \code{runTestSuite}. 
}
\description{
Run package regression tests
}
\details{
Executes the package regression tests, and produces a html report. The test suite is run with the RUnit package.
}
\examples{
\dontrun{
x <- functionMapTests(printTestProtocol = FALSE)
summary(x)
}
}
\author{
Mango Solutions
}
\keyword{debugging}

