% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{functions}
\alias{functions}
\title{All functions of a function map of a package or script}
\usage{
functions(map)
}
\arguments{
\item{map}{Function map.}
}
\description{
All functions of a function map of a package or script
}

