% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{functions_called}
\alias{functions_called}
\title{All functions that called in a package or R script}
\usage{
functions_called(map)
}
\arguments{
\item{map}{Function map.}
}
\value{
Character vector of functions called.
}
\description{
All functions that called in a package or R script
}

