% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denotationUtils.R
\name{get.exported.names}
\alias{get.exported.names}
\title{get.exported.names}
\usage{
get.exported.names(base.path)
}
\arguments{
\item{base.path}{the path pointing to the source of the R package}
}
\value{
names of exported objects
}
\description{
get exported names and patterns from NAMESPACE
}

