% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_base_funcs}
\alias{get_base_funcs}
\title{Names of all functions in the `base` package}
\usage{
get_base_funcs()
}
\value{
Character vector.
}
\description{
Names of all functions in the `base` package
}
\keyword{internal}

