% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces.R
\name{get_exports}
\alias{get_exports}
\title{Get exported functions from a package, from `NAMESPACE`}
\usage{
get_exports(path, funcs)
}
\arguments{
\item{path}{Path a package root.}

\item{funcs}{Character vector of function names, parsed
from the source files. This is needed in case there is an
export pattern in the `NAMESPACE` file}
}
\value{
Character vector of exported functions.
}
\description{
The result includes S3 methods, in the usual
`generic.method` notation.
}
\keyword{internal}

