% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRscript.R
\name{get_funcs_from_r_script}
\alias{get_funcs_from_r_script}
\title{Extract all functions from an R script}
\usage{
get_funcs_from_r_script(rfile)
}
\arguments{
\item{rfile}{The .R input file.}
}
\value{
Named list of function objects, they also include
  the source code, i.e. they have `srcref` attributes.
}
\description{
Reads the file into a temporary environment, and checks whether
the objects in this environment are functions.
}

