% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces.R
\name{get_imports}
\alias{get_imports}
\title{Table of imported functions}
\usage{
get_imports(path)
}
\arguments{
\item{path}{Root folder of the R package.}
}
\value{
Table of imports
}
\description{
First column is the name of the function, second is the
name of the package it is imported from.
}

