% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denotationUtils.R
\name{guess.s3.from.dir}
\alias{guess.s3.from.dir}
\title{guess.s3.from.dir}
\usage{
guess.s3.from.dir(base.path)
}
\arguments{
\item{base.path}{the path which we want to check}
}
\value{
data.frame of S3 methods
}
\description{
scan a package folder to find if there is any S3 methods
}
\details{
It's basically difficult task to determine if a function is S3.
There are two situations can be discovered by this function.
\enumerate{
\item When the given path is a package, hence the S3method definition are available in the NAMESPACE
\item Some new registered S3 generic method which defined using \code{UseMethod}
}
}

