% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denotationUtils.R
\name{guess.s3.from.names}
\alias{guess.s3.from.names}
\title{guess.s3.from.names}
\usage{
guess.s3.from.names(n,
  knownS3generic = tools:::.get_S3_generics_as_seen_from_package("", F))
}
\arguments{
\item{n}{network of just a vector of names}

\item{knownS3generic}{names of known S3 generic methods}
}
\value{
data.frame describing the vertices
}
\description{
With provided knowlege of possible S3 generic names, return a data.frame describing the category of a vector of function names
Typically used when the target is not a package, do not have a NAMESPACE
}

