% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denotationUtils.R
\name{guess.s4.from.dir}
\alias{guess.s4.from.dir}
\title{guess.s4.from.dir}
\usage{
guess.s4.from.dir(base.path)
}
\arguments{
\item{base.path}{path pointing to the base of a R package}
}
\value{
data.frame of S4 classes and methods
}
\description{
scan a package folder (NAMESPACE file) to get exported S4 class and methods
}

