% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{hits.rank}
\alias{hits.rank}
\title{hubness and authority}
\usage{
hits.rank(net, tol = 1e-05)
}
\arguments{
\item{net}{network object or matrix}

\item{tol}{tolerance for iteration}
}
\value{
h and a vector
}
\description{
hub \code{h} means how important the caller is.
}
\details{
authority \code{a} means how important the callee is.
}

