% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_list}
\alias{in_list}
\title{Check if an object is in a list}
\usage{
in_list(elem, list)
}
\arguments{
\item{elem}{The object.}

\item{list}{The list.}
}
\value{
Logical scalar
}
\description{
Check if an object is in a list
}
\keyword{internal}

