% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{interconnected}
\alias{interconnected}
\title{inter connected edgelists}
\usage{
interconnected(el)
}
\arguments{
\item{el}{edgelist}
}
\value{
data.frame
}
\description{
Subset the edgelist to a smaller one that the callees are defined in the same package.
}
\details{
For each defined function(tails), add attributes of text to record other callees which are not defined in this package.
If a caller only calls functions outside this package, a name 'outpackage_functions' will be used to replace the actual name.
Weights are summarized accordingly.

Note: after interconnected, the inpackage tagging by \code{\link{network.from.edgelist}} might be incorrect.
}

