% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denotationUtils.R
\name{is.sas.edgelist}
\alias{is.sas.edgelist}
\title{is.sas.edgelist}
\usage{
is.sas.edgelist(x)
}
\arguments{
\item{x}{edgelist object}
}
\value{
TRUE or FALSE
}
\description{
edgelist can be either from rpackage or SAS scripts
Only difference is the last column is 'rfile' or 'sasfile'
}

