% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{isolated.vertexes}
\alias{isolated.vertexes}
\title{isolated.vertexes}
\usage{
isolated.vertexes(n, need.plot = FALSE)
}
\arguments{
\item{n}{network object}
}
\value{
isolated vertexes names
}
\description{
return isolated.vertexes of a network object
}

