% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{isolates}
\alias{isolates}
\title{Vertices that cannot be found from the a set of source vertices}
\usage{
isolates(graph, sources)
}
\arguments{
\item{graph}{The input graph, an adjacency list.}

\item{sources}{The names of the source vertices.}
}
\value{
A character vector of unreachable vertex names.
}
\description{
Vertices that cannot be found from the a set of source vertices
}

