% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_find.R
\name{load_dependencies}
\alias{load_dependencies}
\title{Load all dependencies of a package, and the package itself, quietly}
\usage{
load_dependencies(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
Attached package namespace, or a `try-error` object
  if an error happened.
}
\description{
Load all dependencies of a package, and the package itself, quietly
}
\keyword{internal}

