% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-exprs.R
\name{make_func_from_exprs}
\alias{make_func_from_exprs}
\title{Create a function from a list of expressions}
\usage{
make_func_from_exprs(exprs)
}
\arguments{
\item{exprs}{The list of expressions to put into the
function body.}
}
\value{
A function.
}
\description{
We do this, so that `codetools::findGlobals` can be
called on the function.
}
\keyword{internal}

