% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_r_folder}
\alias{map_r_folder}
\title{Map the function calls for a folder of R scripts}
\usage{
map_r_folder(rpath, rfilepattern = "\\\\.[R|r]$", include_base = FALSE)
}
\arguments{
\item{rpath}{Character vector of folders and files.
Files are taken as given, folders are listed for files
matching `rfilepattern`.}

\item{rfilepattern}{Pattern of file names to analyse. By
default files with `.R` and `.r` extensions are used.}

\item{include_base}{Whether to include functions from the
`base` package.}
}
\value{
A function_map object.
}
\description{
Map the function calls for a folder of R scripts
}

