% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_r_package}
\alias{map_r_package}
\title{Map an R package}
\usage{
map_r_package(path, include_base = FALSE)
}
\arguments{
\item{include_base}{Whether to include functions from the
\code{base} package.}

\item{rpath}{Name of a source R package, or a folder of an
R package.}
}
\description{
Map an R package
}

