% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_r_script}
\alias{map_r_script}
\title{Map the function calls for an R script}
\usage{
map_r_script(rfile, include_base = FALSE)
}
\arguments{
\item{rfile}{The .R input file, or a connection to read from.}

\item{include_base}{Whether to include calls to base functions
in the output.}
}
\value{
A function_map object.
}
\description{
Map the function calls for an R script
}

