% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRscript.R
\name{match.eval.pattern}
\alias{match.eval.pattern}
\title{match.eval.pattern}
\usage{
match.eval.pattern(e)
}
\arguments{
\item{e}{expression}
}
\value{
\code{TRUE} or \code{FALSE} and matched eval calls
}
\description{
Helper function for \code{\link{analyse.eval.call.pattern}} for recursively matching the eval pattern
}
\examples{
\dontrun{
   match.eval.pattern( body(model.frame.default) )  }
}

