% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{network.from.edgelist}
\alias{network.from.edgelist}
\title{network.from.edgelist}
\usage{
network.from.edgelist(edgelist)
}
\arguments{
\item{edgelist}{the edgelist returned by \code{\link{edgelist.from.rpackage}}}
}
\value{
network object with vertex attributes 
}
\description{
Convert edgelist to network, also denotate the attributes as S3, S4 for the vertices
}
\examples{
\dontrun{
 el = edgelist.from.rpackage('/experiment/RNONMEM/importing/RNMImport')
 el
 nn = network.from.edgelist(el)
 nn
 plot(eForce(nn, use.network.attr=TRUE))
 el2 = interconnected(el)
 el2
 attr(el2,'vertex.other.call')[1:10]
 n2 = network.from.edgelist(el2)
 (n2 \%v\% 'details')[1:10]
 plot(eForce(n2, use.network.attr=TRUE))
 n2 \%v\% 'value' <- ceiling(rank(page.rank(n2)))
 plot(eForce(n2, use.network.attr=TRUE))

 network.vertex.names(nn) [ rank(page.rank(nn))>400 ]
  
}
}

