% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{network.from.edgelist.sas}
\alias{network.from.edgelist.sas}
\title{network.from.edgelist.sas}
\usage{
network.from.edgelist.sas(edgelist)
}
\value{
network object
}
\description{
Similar to \code{\link{network.from.edgelist}} but do category denotation base on SAS rather than R
}

