% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRfolder.R
\name{network.from.rpackage}
\alias{network.from.rpackage}
\title{network.from.rpackage}
\usage{
network.from.rpackage(base.path, rfilepattern = "\\\\.[R|r]$")
}
\arguments{
\item{base.path}{path to the R source, assume it to be a package base path}

\item{rfilepattern}{some author may use extension ".[qQ][sS]" other than [Rr], this option can select from those}
}
\value{
network object
}
\description{
Similar to \code{\link{parseRfolder}}, but differences.
\enumerate{
\item \code{parseRfolder} don't assume the path is a root path to a package, but \code{network.from.rpackage} does.
\item \code{parseRfolder} returns a names list of called functions while this one return a network object which can have more attributes
}
}
\examples{
\dontrun{
  n1 <- network.from.rpackage('MASS')
  (n1 \%e\% 'weights')[1:10]
  plot(eForce(n1))
  plot(eForce(n1, use.network.attr=TRUE ))
  plot(eForce(n1 \%s\% which( n1\%v\% 'category' !='outpackage'), use.network.attr=TRUE ))
}
}

