% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASfolder.R
\name{network.from.sascode}
\alias{network.from.sascode}
\title{network.from.sascode}
\usage{
network.from.sascode(sas.path, sas.pattern = "\\\\.[Ss][Aa][Ss]$")
}
\arguments{
\item{sas.path}{basedir to analysis}

\item{sas.pattern}{the extention file name pattern}
}
\value{
network object
}
\description{
A Wrapper for the \code{\link{parseSASfolder}} and \code{\link{createNetwork}}
Also set vertex attribute for the toplevel scripts.
}
\examples{
\dontrun{
  net <- network.from.sascode('.')
  # it will have the atrribute of being a toplevel script or not
  net \%v\% 'toplevel'
}
}

