% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-frame.R
\name{node_data_frame}
\alias{node_data_frame}
\title{Extract vertex (= function) metadata from the call graph}
\usage{
node_data_frame(map)
}
\arguments{
\item{map}{Function map.}
}
\value{
Data frame, one row for each function.
}
\description{
Extract vertex (= function) metadata from the call graph
}
\keyword{internal}

