% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{node_df}
\alias{node_df}
\title{Extract data about the functions in a function map}
\usage{
node_df(map)
}
\arguments{
\item{map}{The function map object.}
}
\value{
A data frame with one row for each function in the map.
Current columns:
  \item{ID}{The name of the function.}
  \item{own}{Whether it is a function in the mapped script, folder or
    package.}
  \item{exported}{Whether it is an exported function of the package.
    For scripts and folders it is always \code{FALSE}.}
}
\description{
Returns a data frame, where each row is a function in the function
map. The data frame includes functions that are in (other) R
packages, and are called from the mapped script, folder or package.
}

