% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_name}
\alias{package_name}
\title{Get the name of the package from `DESCRIPTION`}
\usage{
package_name(path = ".")
}
\arguments{
\item{path}{Path to the package root.}
}
\value{
Package name
}
\description{
Get the name of the package from `DESCRIPTION`
}
\keyword{internal}

