% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{page.rank}
\alias{page.rank}
\title{page.rank}
\usage{
page.rank(net, beta = 0.85, tol = 1e-05, topic.ind)
}
\arguments{
\item{net}{network object or just a matrix}

\item{beta}{taxation parameter}

\item{tol}{tolerance}

\item{topic.ind}{restric target teleport position}
}
\value{
page rank vector
}
\description{
page.rank
}
\examples{
\dontrun{
     (r <- page.rank(net))
     page.rank(net, topic.ind = net \%v\% 'category' != 'outpackage')
     net \%v\% 'value' <- as.vector(r) * 5000
     plot(eForce(net, use.network.attr=TRUE))
}
}

