% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_collate}
\alias{parse_collate}
\title{Parse the `Collate` field of package `DESCRIPTION`}
\usage{
parse_collate(str)
}
\arguments{
\item{str}{The description string.}
}
\value{
Character vector, the parsed file names.
}
\description{
Parse the `Collate` field of package `DESCRIPTION`
}
\keyword{internal}

