% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_find.R
\name{parse_depends}
\alias{parse_depends}
\title{Parse a DESCRIPTION Depends line}
\usage{
parse_depends(pkg)
}
\arguments{
\item{pkg}{Name of the package. Must be installed.}
}
\value{
Character vector of the names of depended packages.
}
\description{
Parse a DESCRIPTION Depends line
}
\keyword{internal}

