% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRfolder.R
\name{parse_r_folder}
\alias{parse_r_folder}
\title{Parse one or more folders of R scripts}
\usage{
parse_r_folder(rpath, rfilepattern = "\\\\.[R|r]$", include_base = FALSE,
  multiples = FALSE)
}
\arguments{
\item{rpath}{Character vector of folders and files.
Files are taken as given, folders are listed for files
matching `rfilepattern`.}

\item{rfilepattern}{Pattern of file names to analyse. By
default files with `.R` and `.r` extensions are used.}

\item{include_base}{Whether to include functions from the
`base` package.}

\item{multiples}{Whether to include functions as many times as
they are called.}
}
\value{
A named list with one entry for each analyzed functions.
  Each entry contains the names of the functions called.
}
\description{
Parse one or more folders of R scripts
}

