% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctionMap.R
\name{parse_roxygen_export}
\alias{parse_roxygen_export}
\title{Parse Rd docs get function export information}
\usage{
parse_roxygen_export(packagedir)
}
\arguments{
\item{packagedir}{path to root of package}
}
\value{
a data.frame summary for exported (as you documented in Rd by export)
}
\description{
Parse Rd docs get function export information
}

