% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRfolder.R
\name{parseRfolder}
\alias{parseRfolder}
\title{Parse the R scripts in a folder and return the function structure}
\usage{
parseRfolder(rpath, rfilepattern = "\\\\.[R|r]$", returnfilename = FALSE)
}
\arguments{
\item{rpath}{The path of the R script folders or files.}

\item{rfilepattern}{An optional regular expression. Only file names which match the regular expression will be parsed.}

\item{returnfilename}{TODO}
}
\value{
A list of the functions structure.
}
\description{
Parse the R scripts in a folder and return the function structure
}
\examples{
\dontrun{
rpath <- system.file("examples", "R", package = "functionMap")
parseRfolder(rpath)
}

}
\author{
Mango Solutions
}

