% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRscript.R
\name{parseRscript}
\alias{parseRscript}
\title{Parse the R script and return the function structure.}
\usage{
parseRscript(rfile,
  analyse.advance.pattern = getOption("analyse.advance.pattern"))
}
\arguments{
\item{rfile}{Path of the R script file.}

\item{analyse.advance.pattern}{any extra pattern we should also analyse and include}
}
\value{
A list of all the functions included in this script. Each component contains the names of the functions were called by this function. 
}
\description{
Parse the R script and return the function structure
}
\examples{
\dontrun{
rfile <- system.file("examples", "R", "func.R", package = "functionMap")
parseRscript(rfile)
}

}
\author{
Mango Solutions
}

