% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpS4Methods.R
\name{parseS4fromNs}
\alias{parseS4fromNs}
\title{parseS4fromNs}
\usage{
parseS4fromNs(...)
}
\arguments{
\item{...}{parameters passed to \code{\link{dumpS4Generic}}}
}
\value{
a parsed object from \code{parseRscript}
}
\description{
wrapper for \code{\link{dumpS4Generic}} and \code{\link{parseRscript}}
}
\examples{
\dontrun{
## Matrix is a very complex S4 package
parseS4fromNs('Matrix')

ordinary.functions <- parseRfolder(
  system.file("examples", "R", package = "functionMap")
)

## if there are generics definition in .GlobalEnv, you can omit it
fmap_dir <- system.file("examples", "R", package = "functionMap")
for (i in list.files(fmap_dir, full.names=TRUE, pattern='*.R')) {
  source(i)
}
parseS4fromNs()

all.funs <- c(ordinary.functions, parseS4fromNs())
   
nets <- createNetwork(all.funs)
plot(eForce(nets))
}
}

