% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASfolder.R
\name{parseSASfolder}
\alias{parseSASfolder}
\title{parseSASfolder}
\usage{
parseSASfolder(sas.path, sas.pattern = "\\\\.[Ss][Aa][Ss]$")
}
\arguments{
\item{sas.path}{the toplevel path containing SAS scripts}

\item{sas.pattern}{the extension pattern for SAS scripts}
}
\value{
list of contained proc and macros
}
\description{
similar to parse_r_folder, but deal with SAS code
}
\examples{
\dontrun{
   nn <- createNetwork(parseSASfolder('.'))
   plotFunctionMap(nn)
   plot(eForce(nn))
}
}

