% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASscript.R
\name{parseSASscript}
\alias{parseSASscript}
\title{parseSASscript}
\usage{
parseSASscript(sas.script, user.macros.list = sub("\\\\.sas$", "",
  casefold(basename(list.files(pattern = "\\\\.[Ss][Aa][Ss]$", recursive =
  TRUE)))), output.format = "list")
}
\arguments{
\item{sas.script}{script name}

\item{user.macros.list}{the lookup table of user defined macros, if \code{sas.script} has used any of \code{user.macros.list}, it will be reported in the output}

\item{output.format}{if it's \code{list} then return a single character including used proc name and user macro name; or return a data.frame}
}
\value{
a character of user proc and macro if \code{output.format} is \code{list}, or return  a data.frame of script name, number of lines, times of proc and data step used and user macro used
}
\description{
extract lines, used proc, data step and user macro information from SAS script
}
\examples{
\dontrun{
 setwd('SASScriptToplevel')
 parseSASscript('06_Model build.sas', output.format='list')
 parseSASscript('03_ CREATE CSV FOR CLASSING.sas', output.format='data.frame')
 parseSASscript('Macros/ModelCode.sas')
}
}

