% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctionMap.R
\name{plotFunctionMap.package}
\alias{plotFunctionMap.package}
\title{wrapper for plotFunctionMap}
\usage{
plotFunctionMap.package(packagedir, relabel = TRUE, do.plot = TRUE, ...)
}
\arguments{
\item{packagedir}{path to root of package, also should be the name of package}

\item{relabel}{re-label to numbers, and add a legend}

\item{do.plot}{only run this function as a wrapper to create network or actual draw the function map}

\item{...}{Extra arguments to pass to \code{plot}.}
}
\description{
If the packagedir has Rd docs, it will create colorful plot for vertices
}
\examples{
\dontrun{
plotFunctionMap.package('/src/ggplot2', width=10, height=10,
  displayisolates=TRUE, displaylabels=FALSE)
plotFunctionMap.package('/src/plyr', width=10, height=10,
  displayisolates=TRUE, label.cex=0.2)
plotFunctionMap.package('/src/foreach', width=10, height=10,
  displayisolates=TRUE, label.cex=0.2)

# simply use this function to create network only
net. = plotFunctionMap.package('/local/maturity_temp_dir/src/Matrix', do.plot=FALSE)
} 
}

