% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctionMap.R
\name{plotFunctionMap}
\alias{plotFunctionMap}
\title{Plot the network of functions}
\usage{
plotFunctionMap(networkobj, displayisolates = FALSE, displaylabels = TRUE,
  label.cex = 1, arrowhead.cex = 0.8, edge.lwd = 0.1,
  vertex.col = rep(2, networkobj$gal$n), pdffile, width = 5, height = 5,
  ...)
}
\arguments{
\item{networkobj}{The network object which can be produced by \code{\link{createNetwork}}.}

\item{displayisolates}{Should isolates be displayed?}

\item{displaylabels}{Should vertex labels be displayed?}

\item{label.cex}{Character expansion factor for label text.}

\item{arrowhead.cex}{Expansion factor for edge arrowheads.}

\item{edge.lwd}{line width scale for edges.}

\item{vertex.col}{color for vertices; may be given as a vector or a vertex attribute name, if vertices are to be of different colors.}

\item{pdffile}{The file name if output PDF file.}

\item{width}{Width of PDF.}

\item{height}{Height of PDF.}

\item{...}{other arguments e.g. \code{main} which is title, passed to \code{plot.network.default}}
}
\value{
\code{NULL} .
}
\description{
Plot the network of functions
}
\examples{
\dontrun{
lfun <- parseRfolder(system.file("examples", "R", package = "functionMap"))
n1 <- createNetwork(lfun)
plotFunctionMap(n1)
}
}
\author{
Mango Solutions
}

