% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.function_map}
\alias{print.function_map}
\alias{print.function_map_rfile}
\alias{print.function_map_rfolder}
\alias{print.function_map_rpackage}
\title{Print method for a function map object.}
\usage{
\method{print}{function_map}(x, ...)

\method{print}{function_map_rfile}(x, ...)

\method{print}{function_map_rfolder}(x, ...)

\method{print}{function_map_rpackage}(x, ...)
}
\arguments{
\item{x}{Function map to print}

\item{...}{Additional arguments, ignored currently.}
}
\value{
Printed object, invisibly.

Printed object, invisibly.

Printed object, invisibly.

Printed object, invisibly.
}
\description{
The object can be the map of an R file, a folder containing
R files, or an R package. It prints an adjacency list, nicely
formatted.
}

