% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_package_files}
\alias{r_package_files}
\title{Get all source files of a package, in the right order}
\usage{
r_package_files(path)
}
\description{
It uses the `Collate` entry in the `DESCRIPTION` file,
if there is one. Otherwise the order is alphabetical.
}

