% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{remove_loops}
\alias{remove_loops}
\title{Remove loops from a graph}
\usage{
remove_loops(graph)
}
\arguments{
\item{graph}{Input graph, as an adjacency list.}
}
\value{
Another graph, with the loop edges removed.
}
\description{
Remove loops from a graph
}

