% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{rev_deps}
\alias{rev_deps}
\title{Reverse dependencies of all functions}
\usage{
rev_deps(map, multiples = FALSE)
}
\arguments{
\item{map}{Function map.}

\item{multiples}{Whether to keep multiple calls from the same
function.}
}
\value{
Named list of character vectors.
}
\description{
Reverse dependencies of all functions
}

