% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_calls}
\alias{s3_calls}
\title{Find the S3 methods of a function, in an environment}
\usage{
s3_calls(funcname, envir, multiples = FALSE)
}
\arguments{
\item{funcname}{Name of the S3 generic.}

\item{envir}{Environment containing functions, we search in this.}

\item{multiples}{Has no effect.}
}
\description{
Typically used to find the S3 methods defined in a package,
for an S3 generic, also defined in the package.
}
\keyword{internal}

