% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_r_folder.R
\name{s4.source.from.rpackage}
\alias{s4.source.from.rpackage}
\title{s4.source.from.rpackage}
\usage{
s4.source.from.rpackage(base.path, rfilepattern = "\\\\.[Rr]$")
}
\arguments{
\item{base.path}{path to the R source, assume it to be a package base path}

\item{rfilepattern}{some author may use extension ".[qQ][sS]" other than [Rr], this option can select from those}
}
\value{
network object
}
\description{
Extract source code from S4 definitions from a R package
}

