% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sankey_plot}
\alias{sankey_plot}
\title{Sankey plot of a function map}
\usage{
sankey_plot(map, ...)
}
\arguments{
\item{map}{Function map.}

\item{...}{Additional arguments are passed to \code{sankey::sankey}.}
}
\value{
Nothing.
}
\description{
Sankey plots usually show the flow of some quantity, but they can also
be used to visualize functions calls between functions.
}
\details{
By default exported functions in a package are colored dark blue, and
packages called from another package are colored green.

Recursive calls are omitted from the plot.
}

