% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.recharts.R
\name{set}
\alias{set}
\title{recharts set fucntion}
\usage{
set(x, optionList = NULL, ...)
}
\arguments{
\item{obj}{recharts object.}
}
\value{
The output list of recharts as a list.
}
\description{
An shell function for setting arguments of the recharts object.
}

