% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASfolder.R
\name{show.connected.components}
\alias{show.connected.components}
\title{show.connected.components}
\usage{
show.connected.components(nn, make.plot = TRUE)
}
\arguments{
\item{nn}{network object}

\item{make.plot}{make the plot for components}
}
\value{
the list components network
}
\description{
display connected components with more than one node from a graph
}
\examples{
\dontrun{
  foo <- parseSASfolder('.')
  nn  <- createNetwork(foo)
  show.connected.components(nn)
}
}

