% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_trim}
\alias{str_trim}
\title{Trim leading and trailing whitespace from a character vector}
\usage{
str_trim(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Trimmed character vector.
}
\description{
Trim leading and trailing whitespace from a character vector
}
\keyword{internal}

