% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSASfolder.R
\name{toplevel.sas.structure}
\alias{toplevel.sas.structure}
\title{toplevel.sas.structure}
\usage{
toplevel.sas.structure(sas.path, sas.pattern = "\\\\.[Ss][Aa][Ss]$",
  output.file = "network.map.pdf")
}
\arguments{
\item{sas.path}{the path to which containing SAS scripts, assuming the SAS scripts under \code{sas.path} are the main scripts}

\item{sas.pattern}{which pattern to including in analysis as SAS scripts}

\item{output.file}{the output pdf file which displays the structures}
}
\value{
list of network of toplevel scripts
}
\description{
Analysis and visualize toplevel SAS scripts
}

