% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils.R
\name{topo.sort}
\alias{topo.sort}
\title{topo.sort}
\usage{
topo.sort(n, need.rev = TRUE, direction = "forward")
}
\arguments{
\item{n}{network}

\item{need.rev}{compress from left or from right (align to left or align to right)}

\item{direction}{which direction}
}
\value{
compressed (topological) order
}
\description{
compressed topological sort
}

