% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{topo_sort}
\alias{topo_sort}
\title{Topological sort of a graph}
\usage{
topo_sort(graph)
}
\arguments{
\item{graph}{Input graph as an adjacency list.}
}
\value{
Character vector of vertex names in
  topological order.
}
\description{
Topological sort of a graph
}

