% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{twist_graph}
\alias{twist_graph}
\title{Change the direction of each edge to the opposite in a graph}
\usage{
twist_graph(graph)
}
\arguments{
\item{graph}{The input graph, an adjacency list.}
}
\value{
The output graph
}
\description{
I.e. from an in-adjacency list create an out-adjacency list,
and vice versa.
}
\keyword{internal}

