% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_find.R
\name{unload_dependencies}
\alias{unload_dependencies}
\title{Unload all dependencies of a package}
\usage{
unload_dependencies(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\description{
Unload all dependencies of a package
}
\keyword{internal}

