% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{unused_functions}
\alias{unused_functions}
\title{Functions that are never used in a package}
\usage{
unused_functions(map)
}
\arguments{
\item{map}{Function map.}
}
\description{
These are the functions that are never called by an exported
function, neither directly, nor indirectly.
}

