% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces.R
\name{where}
\alias{where}
\title{Try to find where some called functions live}
\usage{
where(map)
}
\arguments{
\item{map}{Function map object.}
}
\value{
Data frame with columns:
  \item{func}{Function name.}
  \item{place}{Name of the package it is coming from.
    Empty string means it is in the mapped package. NA means that
    we don't know where it is coming from.}
}
\description{
Try to find where some called functions live
}
\details{
\enumerate{
  \item They might be in the mapped package.
  \item They might be explicit calls to a package using `::`.
  \item They might be functions imported by name or in a package
    that was imported by a whole.
  \item They might be in a base package that is attached by default.
}
}
\keyword{internal}

