% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_find.R
\name{with_package}
\alias{with_package}
\title{Evaluate an expression with a package loaded}
\usage{
with_package(path, expr)
}
\arguments{
\item{path}{Path to the package directory and tarball to load.}

\item{expr}{Expression to evaluate.}
}
\value{
Value of the evaluated expression.
}
\description{
The package will be installed into a temporary directory.
Then all required dependencies will be loaded, and the package
will be loaded as well.
}
\details{
After the expression is evaluated, all dependencies and the package
itself will be unloaded, and the temporary library directory
will be deleted. The library path will be also restored.
}
\keyword{internal}

