% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRscript.R
\name{convertToCharacter}
\alias{convertToCharacter}
\title{convertToCharacter}
\usage{
convertToCharacter(L)
}
\arguments{
\item{L}{return list of \code{analyse.do.call.pattern}}
}
\value{
character list representing input \code{L}
}
\description{
Helper function for \code{\link{analyse.do.call.pattern}},
which convert return list of \code{analyse.do.call.pattern} to character vector
}

