% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{create_function_map}
\alias{create_function_map}
\title{Create a function_map object from a parsed result}
\usage{
create_function_map(data, package = NULL, rfile = NULL, rpath = NULL,
  rfilepattern = NULL, include_base = NULL, class = NULL,
  active_find = TRUE)
}
\arguments{
\item{data}{A named list. Names are function names,
list entries are the functions they call. Make sure you include
multiplicity.}

\item{package}{Name of the R package, or \code{NULL} if not an R
package.}

\item{rfile}{Name of the R script, or \code{NULL} if not an R script.}

\item{rpath}{Path to the R folder, or \code{NULL} is not an R folder.}

\item{rfilepattern}{Pattern for R files in the folder, or \code{NULL}
if not an R folder.}

\item{include_base}{Whether calls to base functions are included.}

\item{class}{What class to set on the result, in addition to
\code{function_map}.}

\item{active_find}{Whether to try to find call endpoints actively.}
}
\value{
A function_map object.
}
\description{
Create a function_map object from a parsed result
}
\keyword{internal}

