% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-call.R
\name{do_call_globals}
\alias{do_call_globals}
\title{Find the global functions in \code{do.call} patterns}
\usage{
do_call_globals(fun, multiples = FALSE)
}
\arguments{
\item{fun}{A function object.}

\item{multiples}{Whether to include functions as many
times as they are called.}
}
\value{
Character vector of globals from \code{do.call} calls.
Note that this only includes globals that are speficied
}
\description{
Find the global functions in \code{do.call} patterns
}

