% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{edge_df}
\alias{edge_df}
\title{Extract data about the edges (function calls) in a function map}
\usage{
edge_df(map)
}
\arguments{
\item{map}{The function map.}
}
\value{
A data frame with columns:
  \item{from}{Calling function.}
  \item{to}{Called function.}
  \item{weight}{Number of calls from the caller to the callee.}
}
\description{
Extract data about the edges (function calls) in a function map
}

