% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRscript.R
\name{get_global_calls}
\alias{get_global_calls}
\title{Get global function calls from a function}
\usage{
get_global_calls(func, include_base = TRUE, multiples = FALSE)
}
\arguments{
\item{func}{Function object.}

\item{include_base}{Whether to include calls to base functions
in the output.}

\item{multiples}{Whether to keep multiplicity in the result. I.e.
if this argument is \code{TRUE} and \code{func} calls \code{foobar}
twice, then \code{foobar} is included in the result twice.}
}
\value{
A character vector with the names of all functions called.
}
\description{
Note the results are approximate only. R's dynamic nature
does not allow us to always find the global calls reliably.
}
\details{
Calls can be: \itemize{
  \item Direct function calls.
  \item Function calls via \code{do.call}.
  \item Calls to external functions via \code{.C}, \code{.Call}, etc.
}

Internally we use \code{\link[codetools]{findGlobals}} for finding
the global calls and variables.

Note that by default calls to base function are also included in the
result, even of they are primitive functions (e.g. \code{<-},
\code{==}, etc.).
}

