% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRfolder.R
\name{parse_r_folder}
\alias{parse_r_folder}
\title{Parse one or more folders of R scripts}
\usage{
parse_r_folder(rpath, rfilepattern = "\\\\.[R|r]$", include_base = FALSE)
}
\arguments{
\item{rpath}{Character vector of folders and files.
Files are taken as given, folders are listed for files
matching \code{rfilepattern}.}

\item{rfilepattern}{Pattern of file names to analyse. By
default files with \code{.R} and \code{.r} extensions are used.}

\item{include_base}{Whether to include functions from the
\code{base} package.}
}
\value{
A named list with one entry for each analyzed functions.
Each entry contains the names of the functions called.
}
\description{
Parse one or more folders of R scripts
}

