% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{remove_base_functions}
\alias{remove_base_functions}
\title{Remove base function names form a character vector, or a list
of character vectors}
\usage{
remove_base_functions(funcs)
}
\arguments{
\item{funcs}{Character vector, or list of character vectors.}
}
\value{
Filtered character vector, or list of character vectors.
}
\description{
Remove base function names form a character vector, or a list
of character vectors
}
\keyword{internal}

